requirements.txt
fastapi
uvicorn
requests
python-dotenv
scikit-learn
google-api-python-client
google-auth
google-auth-oauthlib
